+++ components/vaadin-date-picker/test/angular2/main.js
@@ -1,7 +1,17 @@
-System.register(['@angular/platform-browser-dynamic', '@angular/core', './app.module'], function(exports_1, context_1) {
+System.register(['@angular/platform-browser-dynamic', '@angular/core', '@vaadin/angular2-polymer'], function(exports_1, context_1) {
     "use strict";
     var __moduleName = context_1 && context_1.id;
-    var platform_browser_dynamic_1, core_1, app_module_1;
+    var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
+        var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
+        if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
+        else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
+        return c > 3 && r && Object.defineProperty(target, key, r), r;
+    };
+    var __metadata = (this && this.__metadata) || function (k, v) {
+        if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+    };
+    var platform_browser_dynamic_1, core_1, angular2_polymer_1, core_2;
+    var TestApp;
     return {
         setters:[
             function (platform_browser_dynamic_1_1) {
@@ -9,14 +19,39 @@ System.register(['@angular/platform-browser-dynamic', '@angular/core', './app.mo
             },
             function (core_1_1) {
                 core_1 = core_1_1;
+                core_2 = core_1_1;
             },
-            function (app_module_1_1) {
-                app_module_1 = app_module_1_1;
+            function (angular2_polymer_1_1) {
+                angular2_polymer_1 = angular2_polymer_1_1;
             }],
         execute: function() {
-            core_1.enableProdMode();
+            core_2.enableProdMode();
+            TestApp = (function () {
+                function TestApp(e, ref) {
+                    this.date = '2000-01-01';
+                    this._host = e.nativeElement;
+                    this._ref = ref;
+                }
+                TestApp.prototype.ngAfterViewInit = function () {
+                    var event = new CustomEvent('readyForTests', { detail: this });
+                    this._host.dispatchEvent(event);
+                };
+                TestApp.prototype.detectChanges = function () {
+                    this._ref.detectChanges();
+                };
+                TestApp = __decorate([
+                    core_1.Component({
+                        selector: 'test-app',
+                        template: "\n    <vaadin-date-picker [(value)]=\"date\" required class=\"bound\"></vaadin-date-picker>\n\n    <form #testForm=\"ngForm\">\n      <vaadin-date-picker ngControl=\"date\" required [_form]=\"testForm\"></vaadin-date-picker>\n    </form>\n    ",
+                        directives: [angular2_polymer_1.PolymerElement('vaadin-date-picker')]
+                    }), 
+                    __metadata('design:paramtypes', [core_1.ElementRef, core_1.ChangeDetectorRef])
+                ], TestApp);
+                return TestApp;
+            }());
+            exports_1("TestApp", TestApp);
             document.body.addEventListener('bootstrap', function () {
-                platform_browser_dynamic_1.platformBrowserDynamic().bootstrapModule(app_module_1.AppModule);
+                platform_browser_dynamic_1.bootstrap(TestApp);
             });
         }
     }